// FrontEnd/src/app/models/Filtros/FiltroPadrao.ts
export class FiltroPadrao {

    constructor() {
        this.Target = new Array<PadraoComboFiltro>();
        this.Regiao = new Array<PadraoComboFiltro>();
        this.Demografico = new Array<PadraoComboFiltro>();
        this.Onda = new Array<PadraoComboFiltro>();
        this.Marca = new Array<PadraoComboFiltro>();
    }

    Target: Array<PadraoComboFiltro>;
    Regiao: Array<PadraoComboFiltro>;
    Demografico: Array<PadraoComboFiltro>;
    Onda: Array<PadraoComboFiltro>;
    Marca: Array<PadraoComboFiltro>;
    CodUser: number;
    CodIdioma: number;
    Sequencia: number;
    Atributo: number;
    TituloGrafico: string;
    ParamTipo:number;
    ParamOndaAtual :number;
    ParamSTB:number;
    ParamBIA:number;
    ParamDenominators:number;
}


export class FiltroPadraoExcel {

    constructor() {
        this.Target = new Array<PadraoComboFiltro>();
        this.Regiao = new Array<PadraoComboFiltro>();
        this.Demografico = new Array<PadraoComboFiltro>();
        this.Onda = new Array<PadraoComboFiltro>();
        this.Marca = new Array<PadraoComboFiltro>();

        this.Onda1 = new Array<PadraoComboFiltro>();
        this.Onda2 = new Array<PadraoComboFiltro>();
        this.Onda3 = new Array<PadraoComboFiltro>();
        this.Onda4 = new Array<PadraoComboFiltro>();

        this.Marca1 = new Array<PadraoComboFiltro>();
        this.Marca2 = new Array<PadraoComboFiltro>();
        this.Marca3 = new Array<PadraoComboFiltro>();
        this.Marca4 = new Array<PadraoComboFiltro>();
        this.Marca5 = new Array<PadraoComboFiltro>();

        this.Marca6 = new Array<PadraoComboFiltro>();
        this.Marca7 = new Array<PadraoComboFiltro>();
        this.Marca8 = new Array<PadraoComboFiltro>();

        
    }

    Target: Array<PadraoComboFiltro>;
    Regiao: Array<PadraoComboFiltro>;
    Demografico: Array<PadraoComboFiltro>;
    Onda: Array<PadraoComboFiltro>;
    Marca: Array<PadraoComboFiltro>;

    Onda1: Array<PadraoComboFiltro>;
    Onda2: Array<PadraoComboFiltro>;
    Onda3: Array<PadraoComboFiltro>;
    Onda4: Array<PadraoComboFiltro>;

    Marca1: Array<PadraoComboFiltro>;
    Marca2: Array<PadraoComboFiltro>;
    Marca3: Array<PadraoComboFiltro>;
    Marca4: Array<PadraoComboFiltro>;
    Marca5: Array<PadraoComboFiltro>;

    Marca6: Array<PadraoComboFiltro>;
    Marca7: Array<PadraoComboFiltro>;
    Marca8: Array<PadraoComboFiltro>;

    CodUser: number;
    CodIdioma: number;

    TituloGrafico: string;

    ParamTipo:number;
    ParamOndaAtual:number;
    ParamBIA :number;

    ParamDenominators:number;
}